extracted <GameConfigField/>
